Take RouteController apart three parts
